// src/api/codegen/schemas/informedTemplateNodeListItemElementsItem.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ShowDoc 导出接口 OpenAPI
 * 根据 ShowDoc 原始页面自动生成的 OpenAPI 3.0 规范。
 * OpenAPI spec version: 1.0.0
 */
import type { InformedTemplateNodeListItemElementsItemWarnListItem } from './informedTemplateNodeListItemElementsItemWarnListItem'

export type InformedTemplateNodeListItemElementsItem = {
  id: string
  element_type_id: string
  code: string
  name: string
  unit: string
  pinyin: string
  definition: string
  sort: number
  value_type_id: string
  value_type: string
  value_format: string
  is_char: number
  is_number: number
  is_date: number
  is_time: number
  is_date_time: number
  is_bool: number
  is_data: number
  min_len: number
  max_len: number
  decimal_digit: number
  can_counted: number
  value_explain: string
  ban_change: boolean
  has_range: boolean
  required: boolean
  has_formula: number
  has_warn: number
  formula: string
  warn_list: InformedTemplateNodeListItemElementsItemWarnListItem[]
}